+++ src/classes/UI5Classes/UI5Parser/UIClass/AbstractUIClass.ts
@@ -33,6 +33,7 @@ export interface IUIProperty extends IName {
 	typeValues: ITypeValue[];
 	visibility: string;
 	description: string;
+	defaultValue?: string;
 }
 export interface IUIAggregation extends IName {
 	type: string;